Blogs/requirement.txt
asgiref==3.2.7
Django==3.0.7
pytz==2020.1
sqlparse==0.3.1
asgiref==3.2.7
Django==3.0.7
django-appconf==1.0.4
django-mediumeditor==1.0.0
pytz==2020.1
sqlparse==0.3.1
